CharacterCreator MASSIVE fixes, UI cleanup
The only thing left in Character creator is to set the max amount of each body part (e.g. there would be up to hair07) so we can go to maximally 7 and if we cross that we go back to 0.

Also when going left in choosing body parts then going to the last image.