Condition change  "effect" type
At first the plan was to use BufferedImage and save/set it to the condition.
Right now Im thinking about new effects class that will store effect images and with this effectId I will be lolading the correct one.
There will also be a need for
sendMagicEffect(position/entity, effectid)