isInSight better usage
In Entity.java instead of having whole code in an if I just return at the beginning of the draw method.
Code looks cleaner thanks to that.

In TileManager moved tileNum inside the if = microoptimization
Also left a note there TO-DO, because Im not going to loop through whole map in the future... 0,0 is inefficient.